Replace std::any with std::shared_ptr
std::any is giving me grief because of its copyability: because std::any
is copyable, we can get in trouble because the `bstring_view` paired
with the any will break if the `any` gets copied and contains a
short std::string with small-string-optimization, which leads the
segfaults and other weird bugs when a small string gets passed in.

This replaces std::any with std::shared_ptr<void>, which will do the
same keep-alive effect but without any copying of the underlying
instance.