Add chunked send interface
This allows you to queue up a callback to be invoked multiple times to
be able to send chunks via callbacks that get automatically invoked as
each chunk finishes being used by the stream buffer.

It also includes an (optional) final callback to invoke when the chunks
are done so that you can know when it is safe to append post-chunk data
to the stream.

See the added test code for an example of both.