Unload the shotgun
`io_result` had a bad bug in it by being *implicitly* convertible to
bool, but that meant code such as:

    auto rv = /*something returning io_result*/;
    if (rv != 0) {
        // Error!
    }

was incredibly unintuitively doing the exact opposite thing you would
expect, because the only comparison the compiler can find there is with
the implicit conversion to bool, thus it becomes:

    if ((bool) rv != (bool) 0)

i.e.

    if ((rv.error_code == 0) != false)

i.e.

    if ((rv.error_code == 0) == true)

i.e.

    if (rv.error_code == 0)

This commit unloads the shotgun by adding `.success()` and `.failure()`
methods that are more descriptive, and removing the `operator bool()`
completely (although it would be fine if explicit, `if (rv.success())`
and `if (rv.failure())` just look nicer than `if (rv)` and `if (!rv)`,
especially when working with C code where this would be inverted from
logic on the raw int return values.

This logic inversion also massively improves a performance bottleneck
where we were exiting the flush streams after a single packet (because
we'd interpret success as a failure).