Move shutdown walk into loop call
We want this to happen before we call `stop()` to initiate an event loop
stop; currently this happens after we've stopped, which could put it
after the event loop has already stopped.

That broke things all to hell, though, because of uvw stupidity, so as a
workaround just disable the walk for now until the upcoming rework of
endpoint.