Add CI; add static builds; add Windows support
- Adds CI for the project.
- Adds static dep builds
- Makes windows build work in mingw (though doesn't appear to *work*
  yet).

This adds a suite of Drone CI builds, and fixes various issues where
things weren't building.

This includes full static builds, via the same -DBUILD_STATIC_DEPS=ON
mode we have in other software.

Finally this adds a Windows build, including lots of ugly unholiness
brought in from hell, including a suitably named in constexpr constant
to help tell when we are in such an environment.