Disable ECN for Windows
Windows ECN seems to be...interesting, and also only supported with a
new enough version of Windows 10.  As such, we simply won't use it for
now.  We can always support it later if the need or desire arise.

The unit tests pass, though the change from delays to synchronized tests
is not merged yet so to get all the tests to pass (on my machine, debug
build), log level must be set to info or higher for the tests to
complete quickly enough.