Simplify dgid calculation
We were calculating the dgid by tracking the last dgid in use, and
including the split/side bits in the last dgid, both of which complicate
the math a little.

This changes it to track the *next* dgid and doesn't track the
split/side bits, which makes it easier to understand.

This also (serendipitously) corrects a minor issue with numbering the
first packet: instead of numbering it as either 0 or 2 (depending on
whether split) we would number it as 4 or 2 (effectively skipping dgid 0
for an unsplit packet).