init gnutls from ed keypair
A GNUTLSCreds can now be initialized from a pair of Ed25519 seed/pubkey.

This enables using libquic in "raw public key" TLS mode, which is
useful/needed for Lokinet.

GNUTLS says it checks that the keys given match one another, but
it does not appear to do so.

This commit includes lokinet-specific ALPN code, as testing it in this
codebase is easier at the moment.  The ALPN specifics and Lokinet
specifics should be separated and the Lokinet specifics should move into
Lokinet.

GNUTLSCreds in raw pubkey mode accepts a verify callback which, given the
other side's public key and requested ALPN will determine whether the
connection should be allowed.

Use case for Lokinet:
  clients send the client ALPN during negotiation,
  relays send the relay ALPN on outbound, and
  relays set both for inbound.
  This is used to determine whether we are being connected to by
  a relay or a client, and check keys / setup lokinet accordingly.

if we need gnutls debug logs, we can set
GNUTLS_DEBUG_LEVEL=level (0-99, >10 is full verbosity it seems)
and gnutls will log to stderr