Add connection established and closed callbacks
Connection established is called when the connection has confirmed the
TLS handshake; server does this on ngtcp2 handshake "complete" after
sending the final packet for the handshake to the client which will then
call ngtcp2 handshake "confirmed".

Connection closed callback is called when a connection closes for any
reason, including that it timed out and was never actually established
(thus acting as a "connection attempt failed" signal as well).

In order to have a connection timeout call connection closed, a
handshake timeout needed to be set in ngtcp2 (the default was
funtionally infinite).  5s should be sufficient, but this can be made
configurable at a later time.

Tests which were confirming connection establishment or failure have
been updated to use the new callbacks.