work around incorrect ngtcp2 wakeup time
For whatever reason, sometimes ngtcp2 gives a time point in the past
when asked when to wake it up; this was resulting in us never waking it.
Normally this is fine, as an incoming packet wakes it and we do't notice,
but if no packets come in, new connections we make can fail to timeout
correctly.  This works around that issue by doing a 0µs timeout if the
timeout would be negative.