Apply warnings in tests
Warning flags were being added to the `internal` target, but that has
private linkage so the test suite doesn't pick it up.  This adds a new
`internal-warnings` target that both `internal` and the test code can
link to (without needing to link tests directly to `internal`, which we
don't want).