Remove ENABLE_PERF_TESTING define
This was added so that dgram speedtest could get the dgid itself to use
it to figure out the packet sequence.  The downside is that it required
a compile-time change in the library.

This removes that compile-time change by changing the dgram speedtest
test to mutate the last packet so that we can identify it without
needing to worry about the dgid.