Support for optional options
Certain options in the current interface in awkward in contexts such as
Python where whether you want an option is dependent on some runtime
value: such as being passed a `datagram=True` keyword argument.

Currently the only way we could do it would be to have two separate
paths for each non-nullable option, i.e. something like:

    std::shared_ptr<Endpoint> e;
    if (want_datagram) {
        e = net.endpoint(addr, opt::whatever{},
            opt::datagram_enabled{});
    } else {
        e = net.endpoint(addr, opt::whatever{});
    }

because there's no way to make a `opt::datagram_enabled` that means
*don't* enable datagrams.

While we *could* add such a feature to datagram_enabled and any other
using-this-turns-it-on option, it seems a lot simpler to add a generic
`std::optional<T>` handle to option parsing so that you can pass a
`nullopt` anywhere you don't want to pass an option, i.e. the above can
become:

    std::optional<opt::datagram_enabled> dgrams;
    if (want_datagrams)
        dgrams.emplace();
    auto e = net.endpoint(addr, opt::whatever{}, dgrams);

and similarly for any other of the templated option types.