Remove dubious double-enable_shared_from_this
The one in connection_interface is used, but the second one in
Connection doesn't appear to be, and seems potentially dangerous: we
have two different shared_pointers inside an Connection instance and it
isn't clear that they are aware of each other.

Removing it affected nothing else, so seems fine to drop.