Make `send(bstring)` less dangerous
Currently we don't allow calling `send` with a std::string or ustring,
but a bstring slips through the cracks via the base method
implementation: calling `send(my_bstring)` ends up calling effectively
`send(bstring_view{my_bstring})`, but this ends up sending garbage
because that `my_bstring` likely doesn't live long enough.

This rearranges the virtual method to be protected so that bstring gets
the same protection as std::string and ustring, avoiding an error-prone
case.

(I encountered this in innocuous-looking code that was doing exactly
this, and ending up sending garbage on the wire).

*With* this change, `send(my_bstring)` no longer compiles, forcing you
to change it to `send(std::move(my_bstring))` or else doing something
with a keep-alive, as appropriate.