Don't fail for sending an empty string
There's nothing specifically *wrong* with sending an empty string: it
just doesn't do anything (just like you can print an empty string to a
file or stdout and have it do nothing).

This was throwing (unexpected) exceptions when I casually sent some data
that happened to be empty, and it felt weird to have to make the
application do a `if (!empty)` everywhere it sends something.