Add stream buffer ownership transfer test case
I wrote this up while investigating a crash on the Python wrapper side.
This passes just fine here (and so it's a python wrapper issue), but
having written it it seems a worthwhile test to have.

(We have other tests that explicitly rely on these things, but not one
that definitely tests just this one issue).