Fail for bad input on request length parser
Currently the logic here continually accumulates a buffer until a `:` is
found, but someone could abuse that by sending a very large stream of
data that never contains a `:` to run the receiver out of memory.  This
change puts a maximum on the length parser to mitigate this.

(There was also a bug here that was checking `.back()` instead of
`npos`)