Convert btstream to use more bstring and bstring_views
There were some unnecessary copies in here, but also using a bstring
matches much better what we're actually sending on the wire,
particularly when interacting with Python.  The internals now use
bstring/bstring_views and the public-facing APIs now take the body as
either a regular string_view or a bstring_view.