Defer libquic object destructions to the event loop
This adds a `.make_shared<T>` to Network and Endpoint that works almost
like std::make_shared<T>, except that it ensures that the destructor
will be invoked in the Network's event loop, which is not necessarily
the thread that releases the last shared_ptr.

This is implemented by a two-pass custom deleter: when the last
shared_ptr is released, rather than immediately triggering destruction
(as in a standard shared_ptr) we instead queue the raw pointer
destruction onto the event loop thread to perform the destruction.

This commit also makes the stream, btstream, and datagram constructors
protected, with Network as a friend class, so that they *have* to go
through net.make_shared (because std::make_shared will produce a
shared_ptr that destructs in the wrong place).