(squash me) message: nicer solution to copy/move issues
Updating the string_views was a bit hairy when, essentially, what we
need are offset/sizes of the substr within `data`.  The changes the
implementation to just store offset/size pairs, rather than
string_views, which simplifies things a lot as we can now just use the
default copy/move constructors again.