Make bt request stream timeouts work
BTRequestStream::check_timeouts was't being called; this wires it up via
the existing Endpoint interval, and makes it generic (so other custom
Stream types could use it as well).

Also fixes an issue in the `message` constructor for a timeout, which
doesn't have a body, but was throwing as a result.