BTRequestStream: generic handler & exception catching
This adds the ability to construct a BTRequestStream with a generic
command handler so that if you have a non-trivial set of commands you
don't have to incur a potentially large instantiation cost of
constructing a bunch of lambda for `register_command` calls that may
never be used.  Instead you can now provide a single handler that gets
called for any incoming command.

This also adds the ability to throw a new `no_such_endpoint` exception
so that such a handler can signal that an endpoint doesn't exist.

Since this necessitated adding an exception handler anyway, this also
catches generic exceptions and wraps them with a generic error-producing
response function.  Previously an uncaught exception throw from a
handler would crash the program.