Rename register_command to register_handler
"command" is a bit of a misnomer from an early idea where we were
copying oxenmq's distinction of "requests" and "commands", but
BTRequestStream no longer has that distinction (whether to send/accept a
reply is application-level protocol logic).

This renames it to the better name `register_handler` (and renames
`register_command_fallback` to `register_generic_handler` to match).