Add flag to version the .so
When packaging for the system we need liboxenquic.so.x.y so that we can
have different major/minor versions of the library installed at the same
time.

This means we have to be careful with version numbers: any change to the
API must increment the major or minor version; increments to .z can only
include fixed logic, but no API changes.