Fix segfault from race condition
The TLS key verify callback here can get called *before* the `client_ci`
assignment later in the test block has finished (I see it about 1/5 of
the time on a release build, and I don't think I ver have in a debug
build).  If that happens the `client_ci->open_stream` is dereferencing a
null pointer and segfault.

This fixes it via a promise/future synchronization.