Remove NDEBUG guards around datagram test suite debug fields
Currently these introduce a ODR problem when trying to compile an
application with (or without) NDEBUG against a libquic that was compiled
without (or with) NDEBUG because the library and the code will have (or
expect) different things: the library's `Connection` class would have a
different size than the application things it should have (because of
presence or non-presence of the the debug_interface member), and that
isn't allowed.  (The NDEBUG around the destructors is a similar problem:
an application making a debug build but using a release libquic library
would get linking errors because the release libquic wasn't built with
the destructors existing).

This fixes the issues by making the fields always available, but makes
them non-atomic (so we can avoid atomic overhead) and controls/accesses
them in a thread-safe way via the TestHelper friend class.