Fix delete_connection not getting fired when closing
When we explicitly close a connection we were only calling
`delete_connection` if we *failed* to send the close packet, so this
fixes it.

But the fix also revealed that the callback after sending a packet
wasn't called properly if the packet sent successfully right away
(despite what the doc string says), so this fixes that too to make it
work.

(Tests not included here, but a later commit catches/tests this
problem).