Add configurable idle timeout; change default to 30s
Adds an option passable to `connect`/`listen` to specify a max idle
timeout for outgoing connection/incoming connections.

This also changes the default to 30s instead of 5min because 5min is
long enough that it could cause problems (because internet middle boxes
reportedly start dropping things after 30s or so of no activity).  Thus
a 30s default seems safer, and code that knows it is safe to go longer
can opt-in to larger values.

Also fixes an issue identified by the added test where the close
callback was just getting 0 when we land on it via `drop_connection`; it
now gets an appropriate error code passed through.