BTRequestStream: fire pending callbacks as timeouts when closing
This updates BTRequestStream so that, if the connection (or,
technically, just the stream) closes, we fire off any pending response
callbacks as timeouts.

This is quite useful for application code because, with this change, it
can know that it will get the reply callback fired one way or another
(unless using close_quietly, which explicitly bypasses callbacks).

Without this applications would have to juggle and coordinate calls
across both the close callback and the timeout callbacks to check the
two ways a call could fail.

This also wraps the various callback invocations in `try`s that catch
and log (but otherwise continue) in case application code throws.