Make oxen/quic/version.hpp header less annoying
This makes the version value part of the library instead of part of the
header, because having it in the header makes it annoying to work with
quic as a submodule in an application -- the "quic/version.hpp" include
in quic.hpp isn't found because of the cmake-time generation.

This does the generation in a .cpp file instead; we lose constexpr-ness,
but that doesn't seem like a big deal.