Race hack
There's *still* an elusive race condition in this test (which is perhaps
understandable, since the test itself is specifically designed to test a
network race condition in the first place).

This commit adds a sleep to hopefully resolve it by providing long
enough between the test code and destruction to resolve it.  Fingers
crossed.