Remove from sent_reqs before invoking the callback
If something in the stream callback triggers a connection close then we
could result in a double-call of the callback because, during the call,
the request is still in `sent_reqs` and so a stream closure during that
call would end up in check_timeouts, see that the request is still
there, and call it.

This fixes it by removing the request *before* we invoke the callback so
that if something like that happens it can't double-invoke the callback.