Fix connection dropping during packet processing
Because these were raw pointers, a call to drop a connection in the
middle of processing could leave us with a no-longer-valid Connection
pointer later in the method processing the packet, which crashed.

This changes them to shared pointers to make sure they survive the
duration of the method, even if something causes the Endpoint to drop
them mid-call.