Privatize Endpoint
Endpoint currently has some things exposed that are unsafe to call
whenever you want, so this privatizes much of it.

As part of this, the ngtcp2 callbacks are all moved inside a private
`struct Callbacks` inside connection.cpp that is declared a friend so
that any of the callbacks can still access what they need inside
Endpoint.

Aside from moving things around for this change to keep private/public
things more or less together in the source code (and a couple renames of
functions that used to end in `_cb` but are now inside Callbacks:: where
the `_cb` suffix seemed redundant), there are no logical changes in
this commit.