Properly detect incoming local address
We were previously unilaterally setting the Path's local address to our
bind address on incoming packets.  This, however, breaks connectivity on
multi-IP systems when bound to the any-address for binding: because we
are not specifying an explicit source address the packet goes out on
whatever IP the kernel decides to send it on (which is very likely
whichever it sees as the primary address).

This breaks QUIC connections, though, because you end up with this
scenario:

    - Server has two IPs, 1.2.3.4 and 5.6.7.8.
    - The server's preferred address is 5.6.7.8, and so packets not
      specifically using a source address will be sent from this
      address.
    - Client initiates a connection to server S on IP 1.2.3.4
    - Server responds to the incoming packet, but responds from 5.6.7.8
    - Client rejects the incoming packet from 5.6.7.8 because it is
      expecting something back from 1.2.3.4.

This commit is the first part of fixing this by properly identify the IP
address that we received the packet on (rather than the address we are
bound to) when we are a server for the "local" part of the path.

We only do this when a server because for a client, doing that would
break things because our *outgoing* paths are all constructed using our
bind address, and so we're expecting the return path to match.  (We also
don't have the same constraint for clients because they are allowed to
migrate paths, which means using the any address is fine: it'll be
consistent until it isn't, at which point it migrates to whatever the
new one is).  And if you *really* want a specific IP as a client, you
just bind to it instead of the any address.

The second part, in the following commit, is to properly set the source
address to the one in the path when sending outgoing packets for inbound
connections.

Other changes in this commit:

- moves `Packet` from messages.hpp/cpp into udp.hpp/cpp because it is
  highly specific to and overlapping with the UDP code under quic, while
  `messages.hpp` is more about messages/datagrams on top of quic.

- fixes a bug where we weren't giving a pointer for receiving control
  messages for recvmsg/recvmmsg, which meant we weren't getting any
  control messages, which we needed to receive ECN status.  (We *also*
  need the control messages to get the packet destination address, which
  is why this got noticed and fixed here).