Set packet source address when appropriate
When we are bound to the any-address but our path.local is not the
any-address then we need to tell the OS what source address to use.

This does it, along with a test that fails properly (unfortunately it
requires Linux as I can't see a generic way to make this work in the
test suite as it depends on being able to use multiple IPs from
127.0.0.x in order to have a multi-IP environment but that doesn't work
on Mac/Win).