Fix close hooks on pending streams not firing on connection fail
We were on closing all the *active* streams on a connection failure, but
that means BTRequestStream on *pending* streams would never get called.

This fixes it to execute hooks on *all* streams.

Also moves _conn clearing into Stream, as that seems a better place to
do it (and makes sure it gets caught for all stream, established or
pending).