Remove fmt and oxen logging from headers
Having them in headers breaks packaging because it means we would have
to build packages for both oxen-logging, spdlog, and fmt, which is a
nuissance.

This PR changes the code so that we only use fmt/spdlog/oxen-logging
internal (in src/ and test/) but *don't* use or require it from any
headers.

There is a tiny bit of fmt-dependent code left in quic/format.hpp, but
this is now not included by default and is an optional include an
application with fmt available can include to be able to format quic
types.

For the most part this just moves logs into the cpp and changes a few
formatted strings to work via stl functions, but a few trace logs got
dropped because they seemed unimportant enough to move into the cpp
file.