Make sure sent_requests have a keep-alive
A close at just the wrong time could result in an invalid string_view
still pending on a Stream.

This seems too problematic to have the "no ownership" parameter as a
default argument, so I've made it explicit (and added nullptr to the
many test suites places previously using it implicitly).