Downgrade warning firing on connection timeouts
There are way too many of these in storage serever logs in practice just
from connection testing to endpoints that aren't currently reachable,
and this isn't (necessarily) indicative of a local problem.  This
downgrades it to debug; if the application is doing something here that
should be warned about then it can warn about it in a close handler.