Fix potential crash in close_connection
If the connection has been destroyed between the time close_connection
is called and the `call_soon`ed `_close_connection` happens then the
conn will be a dangling reference.

This fixes it by using a reference id, doing a lookup in the call_soon
and doing nothing if the lookup doesn't find a connection.