Fix Endpoint destruction race condition
It is possible for various events to trigger a deferred endpoint drop or
close after its destruction has been scheduled but before it has fired,
in which case the callbacks would attempt to run post-destruction, and
crash.

This fixes them to capture a weak pointer rather than a hard `this`
pointer so that they can check that the object still exists before
proceeding.