Fix bugs caused by packet delayer
Bug 1: there was a rare race condition in the packet delayer where it
could end up locking a weak pointer (in its own thread) and meanwhile,
on the main thread, the packet delayer shared pointer got destroyed,
leaving the lambda inside the packet delayer itself as the only owner of
the packet delayer.  Since the packet delayer also had its own loop,
that meant it would try to destruct its own Loop from within the loop
thread, which meant a thread was trying to join itself.

This commit fixes that issue by rewriting the packet delayer to always
use its endpoint loop, which is made possible by the previous changes in
this branch.

Bug 2 then appeared where sometimes the 0-RTT tests would randomly fail
to complete in the expected time windows, and it turns out this was
caused by packets getting delivered out of order.  This appears to be an
underlying quirk of libevent where, if the event loop wakes up and
multiple events are due, they are not necessarily processed in the same
order (or, alternatively, they may have been scheduled for the same
microsecond).  In either case, this meant we could end up delivering a
0-RTT packet before the initial packet, which failed and could show up
either as a datagram never arriving, or as a stream getting delayed
because it needed to be retransmitted.

This commit fixes *that* bug by adding queues into packet_delayer so
that packets are always processed in order (and an out-of-order call
just results in some calls finding that their packets have already been
sent and doing nothing).