De-submodule ngtcp2
This removes ngtcp2 as a submodulde and instead moves building it
StaticBuild and makes it otherwise a required dependency (which isn't
that big a deal nowadays as ngtcp2 is commonly available).

Not having it in StaticBuild meant that it got configured before
anything in StaticBuild, and so required various hacks into ngtcp2's
cmake variables to make it working.  Simplying having it in StaticBuild
with proper dependencies is much simpler.