Disable GSO at runtime by default
GSO can fail, depending on the hardware/software: for instance, storage
server packets sent via GSO into a Lokinet tun will fail depending on
kernel version.

This commit makes GSO usage a runtime endpoint option that has to be
opted-in, and is disabled by default.  (And of course opting is does
nothing if GSO isn't supported on the system).