Fix windows static build
ngtcp2, particular on Windows, requires NGTCP2_STATICLIB to be defined
to get the linkage right when ngtcp2 is built statically.  When we
switched to building it externally (rather than as a submodule) that
definition got dropped, and windows mingw linkage broke because the
static library was compiled with it, but our headers don't have it,
leading to a bunch of errors along the lines of:

    /usr/bin/x86_64-w64-mingw32-ld: ../src/libquic.a(connection.cpp.obj):connection.cpp:(.text+0x8d7b): undefined reference to `__imp_ngtcp2_conn_update_pkt_tx_time'
    /usr/bin/x86_64-w64-mingw32-ld: ../src/libquic.a(connection.cpp.obj):connection.cpp:(.text+0x915e): undefined reference to `__imp_ngtcp2_conn_handle_expiry'
    /usr/bin/x86_64-w64-mingw32-ld: ../src/libquic.a(connection.cpp.obj):connection.cpp:(.text+0x91a4): undefined reference to `__imp_ngtcp2_strerror'
    /usr/bin/x86_64-w64-mingw32-ld: ../src/libquic.a(connection.cpp.obj):connection.cpp:(.text+0x97b2): undefined reference to `__imp_ngtcp2_conn_get_streams_bidi_left'
    /usr/bin/x86_64-w64-mingw32-ld: ../src/libquic.a(connection.cpp.obj):connection.cpp:(.text+0xa349): undefined reference to `__imp_ngtcp2_conn_read_pkt_versioned'

According to an answer of
https://stackoverflow.com/questions/10059200/how-do-i-fix-undefined-reference-to-imp
these are caused by a mismatch of `__declspec(dllexport/import)` and and
a lack of `__declspec`.  Without `-DNGTCP2_STATICLIB` we end up trying
to use dllimport when linking, but the static library was built without
dllexport, and so it falls apart.  This puts the `-DNGTCP2_STATICLIB`
back into the target so that the exports in the header match the ones it
was built with.