Add no-data stream notification
This add a `opt::stream_notify` stream construction option that enables
a stream to be opened at both ends even if the stream opener has no data
to send yet.  (Under the hood, this causes an initial 0-length stream
frame to be sent to the other end which is explicitly designed into QUIC
to support such a notification).

The also somewhat refactors stream constructors to take a generic option
list so that the option can be provided.  This change *does* break one
case: you can no longer construct a Stream with an explicit `nullptr`
data callback (to get to the second close callback), but that seems
unused in practice aside from one test suite invocation.  (With this
change you simply omit the data callback argument).

Also incorporates some minor code refactoring to address some minor
warnings and small optimizations.