Coalesce speedtests + small improvements
This combines all the functionalities of
{speedtest,dgram-speed,ping}-{client,server} into a single pair of
speedtest-client and speedtest-server binaries, with `--datagram` and
`--ping` on the client for do a datagram or ping test.

A single server, however, can respond to any of the tests, so you can
just run one server and throw different things at it.

Other changes:

- Adds a SPEEDTEST log category forced to info-level logging so that we
  can log things without needing to put it at critical.

- Removes the hashing and checksumming from stream speedtests (they were
  really only intended for debugging early development stream
  consistency)

- Add ping option to skip N responses for RTT stats, which is useful to
  skip the pings sent out during initial connection handshake.

- Added `--no-ping-log` to skip logging of received pings, which is
  particularly useful when ping flooding.