No-credential connections and 0-RTT fixes
This commit adds the ability for a client to connect to a server without
providing a key, and allows listening servers to choose whether client
keys are required, requested, or omitted.  Required means a client must
provide a key, and it must pass the key verification function; requested
means the server will request a key and the client may provide one: the
key verification function is called with an empty key if the client does
not provide one (so that it can still use the ALPN to reject the
connection, for example, if certain ALPNs require keys and others
don't).  Omitted means the server won't request one at all.

For omitted client pubkeys (either not requested, or requested and not
provided but still allowed), the connection's `remote_key()` will return
an empty span.

Client connections can now use GNUTLSCreds::make_unauthenticated to make
a connection without a pubkey.  They can also omit the TLSCreds from the
connect() call entirely, but that won't allow doing things like enable
0rtt.  This connection with such a missing pubkey will fail if the
server requires pubkeys, but otherwise is allowed.

This deliberately renames `set_key_verify_callback` to
`require_client_keys` (thus breaking code) as calling code needs to
decide whether it wants requires or requests behaviour.

This also fixes closely related issues with 0-RTT: under 0-RTT
resumption, key verification callbacks were never called, and so the
remote_key was never set on the connection object.  This is actually
intentional: a valid 0-RTT ticket already contains the connection data,
and so the initial connection that produced the ticket already passed
verification.  This commit contains fixes to this so that they get
properly reflected in the Connection object, and documents this for the
key verification functions.

Other changes:

- remove pointless `gtls_key` struct and make it back into a simple
  typedef, like it used to be.  The indirection added nothing.

- replace Connection `is_validated()` with `is_handshaked()` and
  `is_handshake_confirmed()`.  Validated isn't used anywhere outside of
  libquic's test suite, and doesn't really work with 0-RTT connections
  that don't re-validate.

- remove `set_draining()` and `set_closing()` -- they were pointless
  private methods that simply set the variable of the same name to
  false.  But anything with private access can already just set those to
  false itself.

- Fixed wrong comment above cert_verify_callback_gnutls about what a
  failure return value looks like.  The function was already returning 1
  rather than a negative (which is actually okay), but this intent is
  now clearer in the comment and the code.

- Removed unused `usage_flags`.  It was already tagged
  `[[maybe_unused]]` but was actually unused and didn't seem to have any
  purpose.