Don't fire close callback on failed incoming connections
If an incoming connection closed before handshaking is completed, we are
currently calling the connection close callback, but that doesn't make
sense because we never fired the connection established callback, and so
the application has never been told about this connection.

This changes is so that upon such failures we don't fire the connection
closed callback and just let it die silently.