Fire conn established earlier for 0-RTT connection
Applications (e.g. lokinet) typically do things in the conn_established
callback, but for a 0-RTT server that receives early stream data or
datagrams, the stream gets created *before* the conn_established
callback, which is unintuitive and counterproductive.

This corrects it so that if we do get early data, we fire the conn
established callback immediately before opening to stream so that
whatever initialization is needed (such as queuing an incoming stream)
can be done safely.