Fix opt::stream_notify not working with rejected 0RTT early data
If early data gets rejected but we had already sent the early 0-byte
stream notify opener then we wouldn't try sending another stream notify
because the `_notify` flag was false.

This fixes it by adding a second `_had_notify` property so that we can
flip _notify back to on if the stream data gets reset by early data
rejection, so that the empty notify frame then still gets properly sent
as a 1RTT stream frame.