Rewrite and expose size + data accumulation
This is useful for custom streams that have a similar need to decode a
prefix-length block of data but aren't actually a BTRequestStream.

This also rewrites the accumulation to be simpler.

Finally this allows (and ignores) an empty BTRequestStream data (e.g.
`0:`).  This was a fatal BTRequestStream error before, but could be
useful, for instance as a no-op to keep a connection alive without a
quic-level keep-alive.