More effecient use of stream buffers
Every packet sent on a connection with stream data has one or more calls
to Stream::pending.  If the stream has lots of user buffers, iterating
them each time pending is called is quite slow, so instead we cache
the index and offset in the user buffers where unsent data starts.