Fix late stream ack causing premature connection close
If a stream ack arrives *after* the stream close callback has fired, the
stream ack callback fires but, because the stream was closed, didn't
find the stream and returned NGTCP2_ERR_CALLBACK_FAILURE.  This,
however, made ngtcp2 drop the connection, which is definitely not right.

This commit changes stream_ack to always return 0 to prevent this from
happening when a remote tried to close a stream (but not the whole
connection) because of a stream-specific failure.