Initialize stream ID to -1; improve max stream data comment
Current stream_id is not initialized and so can be random large values
depending on the memory it was initialized from.  This fixes it to
initialize to -1 until a stream id is assigned so that 1) it has sane
values, and 2) you can use it to tell whether a stream is pending.

Also clarifies comments about what stream limit values actually mean.